files/requirements.txt
discord.py==2.3.2
groq==0.9.0
supabase==2.5.3
python-dotenv==1.0.1
aiohttp==3.9.5
asyncio==3.4.3
